Fix Suprsync failed copy attempts (#337)
* Updates failed-copy-attempts if md5sums don't match after copying

* Fix issue with suprsync query filtering

* Fail cleanly when local files don't exist

* Check if temp path already exists

+++ socs/db/suprsync.py
@@ -206,20 +206,21 @@ class SupRsyncFilesManager:
         if session is None:
             session = self.Session()
 
+        if max_copy_attempts is None:
+            max_copy_attempts = 2**10
+
         query = session.query(SupRsyncFile).filter(
             SupRsyncFile.removed == None,  # noqa: E711
             SupRsyncFile.archive_name == archive_name,
+            SupRsyncFile.failed_copy_attempts < max_copy_attempts
         )
-
-        if max_copy_attempts is not None:
-            query.filter(SupRsyncFile.failed_copy_attempts < max_copy_attempts)
 
         files = []
         for f in query.all():
             if f.local_md5sum != f.remote_md5sum:
                 files.append(f)
             if num_files is not None:
-                if len(files) == num_files:
+                if len(files) >= num_files:
                     break
 
         return files
@@ -331,17 +332,30 @@ class SupRsyncFileHandler:
             # Creates temp directory with remote dir structure of symlinks for
             # rsync to copy.
             file_map = {}
+            remote_paths = []
             with tempfile.TemporaryDirectory() as tmp_dir:
                 self.log.info("Copying files:")
                 for file in files:
                     self.log.info(f"- {file.local_path}")
                     tmp_path = os.path.join(tmp_dir, file.remote_path)
                     os.makedirs(os.path.dirname(tmp_path), exist_ok=True)
+
+                    if not os.path.exists(file.local_path):
+                        self.log.warn("Cannot find file {path}", path=file.local_path)
+                        file.failed_copy_attempts += 1
+                        continue
+
+                    if os.path.exists(tmp_path):
+                        self.log.warn("Temp file {path} already exists!", path=tmp_path)
+                        file.failed_copy_attempts += 1
+                        continue
+
                     os.symlink(file.local_path, tmp_path)
 
                     remote_path = os.path.normpath(
                         os.path.join(self.remote_basedir, file.remote_path)
                     )
+                    remote_paths.append(remote_path)
                     file_map[remote_path] = file
 
                 cmd = ['rsync', '-Lrt']
@@ -354,11 +368,6 @@ class SupRsyncFileHandler:
             for file in files:
                 file.copied = time.time()
 
-            remote_paths = [
-                os.path.join(self.remote_basedir, f.remote_path)
-                for f in files
-            ]
-
             res = self.run_on_remote(['md5sum'] + remote_paths)
             for line in res.stdout.decode().split('\n'):
                 split = line.split()
@@ -369,7 +378,19 @@ class SupRsyncFileHandler:
                     continue
 
                 md5sum, path = line.split()
-                file_map[os.path.normpath(path)].remote_md5sum = md5sum
+                key = os.path.normpath(path)
+                if key in file_map:
+                    file_map[key].remote_md5sum = md5sum
+
+            for file in files:
+                if file.remote_md5sum != file.local_md5sum:
+                    file.failed_copy_attempts += 1
+                    self.log.info(
+                        f"Copy failed for file {file.local_path}! "
+                        f"(copy attempts: {file.failed_copy_attempts})"
+                    )
+                    self.log.info(f"Local md5: {file.local_md5sum}, "
+                                  f"remote_md5: {file.remote_md5sum}")
 
     def delete_files(self, delete_after):
         """
